+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 ### Fixes
 - exit code on on failed modules changed from 0 1o 1
 - changed typo `Deployemnt` to `Deployment` everywhere
+- improved validation and error handling when AccountId is not resolvable from EnvVariable
 
 
 ## v2.1.0 (2022-10-03)